Add files via upload
If a player is unable to bounce the ball then the opponent scores a point and the ball starts to move from center of the screen. Players can control the bat on right hand side using up and down arrow keys and that on left hand side using w or s keys for moving it up or down respectively.